Move ContextMenuEntry (and ContextMenuService) to core-ui
.. because it's a UI element, and this way it'll be able to return an
Image for the Icon
