+++ applications/email/.classpath
@@ -5,5 +5,6 @@
 	<classpathentry kind="src" path="src/main/resources"/>
 	<classpathentry combineaccessrules="false" kind="src" path="/dependencies"/>
 	<classpathentry combineaccessrules="false" kind="src" path="/core-framework"/>
+	<classpathentry combineaccessrules="false" kind="src" path="/core-ui"/>
 	<classpathentry kind="output" path="target/classes"/>
 </classpath>
