+++ applications/email/src/main/java/org/phoebus/applications/email/actions/ContextCreateEmail.java
@@ -5,8 +5,8 @@ import java.util.List;
 
 import org.phoebus.applications.email.EmailApp;
 import org.phoebus.framework.selection.Selection;
-import org.phoebus.framework.spi.ContextMenuEntry;
 import org.phoebus.framework.workbench.ApplicationService;
+import org.phoebus.ui.spi.ContextMenuEntry;
 
 /**
  * A prototype action entry for starting the email application
