+++ applications/pvtable/src/main/java/org/phoebus/applications/pvtable/ContextMenuPVTableLauncher.java
@@ -12,8 +12,8 @@ import java.util.List;
 
 import org.phoebus.core.types.ProcessVariable;
 import org.phoebus.framework.selection.Selection;
-import org.phoebus.framework.spi.ContextMenuEntry;
 import org.phoebus.framework.workbench.ApplicationService;
+import org.phoebus.ui.spi.ContextMenuEntry;
 
 /** Entry for context menues that starts PV Table for selected ProcessVariable
  *
