+++ applications/pvtree/src/main/resources/META-INF/services/org.phoebus.ui.spi.ContextMenuEntry
