+++ build.xml
@@ -7,12 +7,12 @@
 	
   <target name="clean" description="Remove all artifacts">
     <ant target="clean" dir="core/framework"/>
-    <ant target="clean" dir="core/logging"/>
   	<ant target="clean" dir="core/util"/>
   	<ant target="clean" dir="core/vtype"/>
     <ant target="clean" dir="core/pv"/>
-    <ant target="clean" dir="core/types"/>
     <ant target="clean" dir="core/ui"/>
+    <ant target="clean" dir="core/logging"/>
+    <ant target="clean" dir="core/types"/>
     <ant target="clean" dir="applications/greetings"/>
     <ant target="clean" dir="applications/probe"/>
     <ant target="clean" dir="applications/logbook/inmemory"/>
@@ -29,12 +29,12 @@
     <echo message="App classpath: ${toString:app-classpath}"/>
     -->
     <ant dir="core/framework"/>
-    <ant dir="core/logging"/>
     <ant dir="core/util"/>
     <ant dir="core/vtype"/>
     <ant dir="core/pv"/>
-    <ant dir="core/types"/>
     <ant dir="core/ui"/>
+    <ant dir="core/logging"/>
+    <ant dir="core/types"/>
   </target>
 
   <target name="applications" depends="core" description="Build application modules">
